Computação Grafica/image_filter_app/requirements.txt
# Framework web
Flask==2.3.3
Werkzeug==2.3.7

# Processamento de imagens
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2

# Utilitários
python-dotenv==1.0.0